requirements.txt
llvmlite==0.43.0
metalcompute==0.2.5
numba==0.60.0
numpy==2.0.0
opencv-python==4.10.0.84
PyYAML==6.0.2
scipy==1.13.1
tqdm==4.66.4
